requirements.txt
logfire==0.50.1
anthropic==0.3.7
assemblyai==0.17.0
httpx==0.23.3
pyjwt==2.6.0
cryptography==39.0.1
aiofiles==22.1.0
yt-dlp==2023.3.4